rename cellranger reader, put framework in to read dragen metrics
